+++ crates/holochain_p2p/src/spawn/actor.rs
@@ -115,7 +115,7 @@ impl WrapEvtSender {
         dna_hash: DnaHash,
         kitsune_space: Arc<kitsune_p2p::KitsuneSpace>,
         basis_loc: u32,
-        limit: usize,
+        limit: u32,
     ) -> impl Future<Output = HolochainP2pResult<Vec<AgentInfoSigned>>> + 'static + Send {
         timing_trace!(
             {
